Rename model folder to custom
