+++ src/custom/CustomInterfaces.ts
